% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCLUST.R
\name{HCLUST}
\alias{HCLUST}
\title{Hierarchical clustering method}
\usage{
HCLUST(data, k)
}
\arguments{
\item{data}{data as a matrix}

\item{k}{number of clusters}
}
\value{
vector of cluster IDs
}
\description{
Hierarchical clustering method
}
\examples{
HCLUST(as.matrix(c(1,2,3,4,1,2,3,4,5,6,1,2,3,4,5)), 3)
}
