% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPECC.R
\name{SPECC}
\alias{SPECC}
\title{Spectral Clustering}
\usage{
SPECC(data, k, epsilon)
}
\arguments{
\item{data}{data as a matrix}

\item{k}{number of clusters}

\item{epsilon}{distance metric}
}
\value{
vector of cluster IDs
}
\description{
Spectral Clustering
}
\examples{
SPECC(as.matrix(c(1,2,3,4,1,2,3,4,1,2,3,4)), 2, 3)
}
